[Fix]MenuIconを追加
